Load csv at startup. Add some integration tests. Fix some bugs.

+++ src/main/java/jackson/mark/munro/model/Summit.java
@@ -9,7 +9,7 @@ import lombok.EqualsAndHashCode;
 public class Summit {
     @EqualsAndHashCode.Include
     private String name;
-    private int heightInMetres;
+    private double heightInMetres;
     private SummitCategory summitCategory;
     private String gridReference;
 }
